fix assertion under heavy cache usage due to 32-bit overflow (c.f. http://www.wildfiregames.com/forum/index.php?showtopic=13622)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@8317 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/allocators/headerless.cpp
@@ -586,10 +586,10 @@ public:
 	}
 
 private:
-	size_t m_totalAllocatedBlocks, m_totalAllocatedBytes;
-	size_t m_totalDeallocatedBlocks, m_totalDeallocatedBytes;
-	size_t m_currentExtantBlocks, m_currentExtantBytes;
-	size_t m_currentFreeBlocks, m_currentFreeBytes;
+	u64 m_totalAllocatedBlocks, m_totalAllocatedBytes;
+	u64 m_totalDeallocatedBlocks, m_totalDeallocatedBytes;
+	u64 m_currentExtantBlocks, m_currentExtantBytes;
+	u64 m_currentFreeBlocks, m_currentFreeBytes;
 };
